Split scope into a new file
